Separate render logic into separate classes
Adds a Renderable class which is probably closer to something like a
Mesh. It provides support for generating a model-view matrix from the
subclass's model matrix and the provided world view matrix.

Adds a Box class which has all the logic for drawing a box.

The GLApplication now creates 9 boxes to prove out the separated
rendering logic for the boxes. I suspect this would perform poorly for
some large number of boxes. Next step for the box should be to add
support for vertex buffer objects and instances rendering.